refactor(order): OrderStatus 디렉토리 위치 이동(constant) - 리뷰 반영
